% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{matrices}
\alias{matrices}
\alias{A}
\alias{B}
\title{The A matrix}
\usage{
A(lambda, psi)

B(lambda, psi)
}
\arguments{
\item{lambda}{loading matrix of size p*k we have now}

\item{psi}{the variance matrix of size p*p of common factor we have now}
}
\description{
The A matrix
}
